Fixed compile errors on client
